[frontend] move jwt.js to lib
