+++ salz-web/lib/jwt.js
